Rename file to resolve conflicts
